Add evaluation method to CalcModel and implement in CalcCLIApp
The CalcModel was extended to include a method `eval` for evaluating internal computations. A new `eval_internal` private method implements an in-order evaluation of an arithmetic expression contained in a stack of computational elements. These changes enable the calculator to not only build but also evaluate arithmetic expression. Corresponding tests have been implemented in CalcCLIApp.